MDEV-25512 Deadlock between sux_lock::u_x_upgrade() and sux_lock::u_lock()
In the SUX_LOCK_GENERIC implementation, we can remember at most
one pending exclusive lock request. If multiple requests have been
enqueued, the WRITER_WAITING flag will be cleared when the first
waiting writer acquires the exclusive lock.

ssux_lock_low::update_lock(): If WRITER_WAITING is set, wake up
the writer even if the UPDATER flag is set, because the waiting
writer may be in the process of upgrading its U lock to X.

rw_lock::read_unlock(): Also indicate that an X lock waiter must
be woken up if an U lock exists.

This fix may cause unnecessary wake-ups and system calls, but this
is the best that we can do. Ideally we would use the MDEV-25404
composition of sux_lock, but there is no portable way to request
that a non-recursive mutex be created.

Also, we add a unit test for srw_mutex, srw_lock, ssux_lock, sux_lock
so that problems like this can be caught more reliably.